+++ sprite.hpp
@@ -18,8 +18,8 @@ namespace rpg
         // x,x              : the top left corner of the subsprite
         // width, height    : the dimensions of the subsprite
         // origin           : the center part of the subsprite
-        Sprite(const Texture2D& texture, float x, float y,
-               float width, float heigth, Vector2 origin = {0.0f, 0.0f});
+        Sprite(const Texture2D& texture, int x, int y,
+               int width, int heigth, Vector2 origin = {0.0f, 0.0f});
 
         // Renders the sprite for 1 frame
         void DrawSprite(int x, int y, float rotation = 0.0f, Color color = WHITE);
